@@ -1,589 +1,1197 @@
 #include "mqtt_topic_tree.h"
-#include "logger.h"
 #include <algorithm>
 #include <cassert>
 #include <sstream>
+#include "logger.h"
 
 namespace mqtt {
+
+// ============================================================================
+// Helper functions for safe STL operations
+// ============================================================================
+
+// Helper functions removed - using direct implementation in methods
 
 // ============================================================================
 // IntermediateNode 实现
 // ============================================================================
 
-IntermediateNode::IntermediateNode(MQTTAllocator* allocator) 
-    : allocator_(allocator)
-    , children_(TopicTreeAllocator<std::pair<const std::string, std::shared_ptr<TopicTreeNode>>>(allocator))
-    , subscribers_(TopicTreeAllocator<SubscriberInfo>(allocator))
-    , ref_count_(1) {
+IntermediateNode::IntermediateNode(MQTTAllocator* allocator)
+    : allocator_(allocator),
+      children_(TopicTreeAllocator<std::pair<const std::string, std::shared_ptr<TopicTreeNode>>>(
+          allocator)),
+      subscribers_(TopicTreeAllocator<SubscriberInfo>(allocator)),
+      ref_count_(1)
+{
 }
 
-IntermediateNode::~IntermediateNode() {
-}
+IntermediateNode::~IntermediateNode() {}
 
-IntermediateNode* IntermediateNode::clone() const {
-    // 使用allocator分配新的IntermediateNode
-    void* memory = allocator_->allocate(sizeof(IntermediateNode));
-    if (!memory) {
-        throw std::bad_alloc();
+int IntermediateNode::safe_copy_containers(IntermediateNode* dest) const
+{
+  int ret = MQ_SUCCESS;
+
+  // Copy children map
+  if (!children_.empty()) {
+    if (dest->children_.max_size() < children_.size()) {
+      LOG_ERROR("Destination children container too small");
+      return MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
     }
-    
-    IntermediateNode* new_node = new(memory) IntermediateNode(allocator_);
-    new_node->children_ = children_;
-    new_node->subscribers_ = subscribers_;
-    return new_node;
+
+    for (const auto& child_pair : children_) {
+      auto result = dest->children_.insert(child_pair);
+      if (!result.second) {
+        LOG_ERROR("Failed to copy child node");
+        return MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+      }
+    }
+  }
+
+  // Copy subscribers set
+  if (!subscribers_.empty()) {
+    if (dest->subscribers_.max_size() < subscribers_.size()) {
+      LOG_ERROR("Destination subscribers container too small");
+      return MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+    }
+
+    for (const auto& subscriber : subscribers_) {
+      auto result = dest->subscribers_.insert(subscriber);
+      if (!result.second) {
+        LOG_ERROR("Failed to copy subscriber");
+        return MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+      }
+    }
+  }
+
+  return ret;
 }
 
-void IntermediateNode::add_child(const std::string& level, std::shared_ptr<TopicTreeNode> child) {
-    children_[level] = child;
+int IntermediateNode::clone(IntermediateNode*& cloned_node) const
+{
+  int ret = MQ_SUCCESS;
+
+  if (MQ_ISNULL(allocator_)) {
+    LOG_ERROR("Allocator is null");
+    ret = MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+  } else {
+    void* memory = allocator_->allocate(sizeof(IntermediateNode));
+    if (MQ_ISNULL(memory)) {
+      LOG_ERROR("Failed to allocate memory for IntermediateNode");
+      ret = MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+    } else {
+      // Use placement new without try-catch
+      cloned_node = new (memory) IntermediateNode(allocator_);
+      if (MQ_ISNULL(cloned_node)) {
+        LOG_ERROR("Failed to construct IntermediateNode");
+        allocator_->deallocate(memory, sizeof(IntermediateNode));
+        ret = MQ_ERR_TOPIC_TREE_NODE_CREATE;
+      } else {
+        // Safe copy operations with pre-checks
+        ret = safe_copy_containers(cloned_node);
+        if (MQ_FAIL(ret)) {
+          cloned_node->~IntermediateNode();
+          allocator_->deallocate(memory, sizeof(IntermediateNode));
+          cloned_node = nullptr;
+        }
+      }
+    }
+  }
+
+  return ret;
 }
 
-void IntermediateNode::remove_child(const std::string& level) {
-    children_.erase(level);
+int IntermediateNode::add_child(const std::string& level, std::shared_ptr<TopicTreeNode> child)
+{
+  int ret = MQ_SUCCESS;
+
+  if (level.empty()) {
+    LOG_ERROR("Level string cannot be empty");
+    ret = MQ_ERR_TOPIC_TREE_INVALID_TOPIC;
+  } else if (MQ_ISNULL(child)) {
+    LOG_ERROR("Child node cannot be null");
+    ret = MQ_ERR_PARAM_V2;
+  } else {
+    // Pre-check container capacity
+    if (children_.size() >= children_.max_size()) {
+      LOG_ERROR("Children container has reached maximum capacity");
+      ret = MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+    } else {
+      // Safe insert operation
+      children_[level] = child;
+    }
+  }
+
+  return ret;
 }
 
-void IntermediateNode::add_subscriber(const SubscriberInfo& subscriber) {
-    // 如果已存在，更新QoS
+int IntermediateNode::remove_child(const std::string& level)
+{
+  int ret = MQ_SUCCESS;
+
+  if (level.empty()) {
+    LOG_ERROR("Level string cannot be empty");
+    ret = MQ_ERR_TOPIC_TREE_INVALID_TOPIC;
+  } else if (children_.empty()) {
+    LOG_DEBUG("No children to remove");
+    ret = MQ_ERR_NOT_FOUND_V2;
+  } else {
+    auto it = children_.find(level);
+    if (it != children_.end()) {
+      children_.erase(it);
+    } else {
+      LOG_DEBUG("Child level '{}' not found for removal", level);
+      ret = MQ_ERR_NOT_FOUND_V2;
+    }
+  }
+
+  return ret;
+}
+
+int IntermediateNode::add_subscriber(const SubscriberInfo& subscriber)
+{
+  int ret = MQ_SUCCESS;
+
+  if (from_mqtt_string(subscriber.client_id).empty()) {
+    LOG_ERROR("Client ID cannot be empty");
+    ret = MQ_ERR_TOPIC_TREE_INVALID_CLIENT;
+  } else if (subscribers_.size() >= subscribers_.max_size()) {
+    LOG_ERROR("Subscribers container has reached maximum capacity");
+    ret = MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+  } else {
+    // Safe find and erase operation
     auto it = subscribers_.find(subscriber);
     if (it != subscribers_.end()) {
-        subscribers_.erase(it);
+      subscribers_.erase(it);
     }
-    subscribers_.insert(subscriber);
+
+    // Safe insert operation
+    auto result = subscribers_.insert(subscriber);
+    if (!result.second && result.first == subscribers_.end()) {
+      LOG_ERROR("Failed to add subscriber - insert failed");
+      ret = MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+    }
+  }
+
+  return ret;
 }
 
-void IntermediateNode::remove_subscriber(const std::string& client_id) {
+int IntermediateNode::remove_subscriber(const std::string& client_id)
+{
+  int ret = MQ_SUCCESS;
+
+  if (client_id.empty()) {
+    LOG_ERROR("Client ID cannot be empty");
+    ret = MQ_ERR_TOPIC_TREE_INVALID_CLIENT;
+  } else if (subscribers_.empty()) {
+    LOG_DEBUG("No subscribers to remove");
+    ret = MQ_ERR_TOPIC_TREE_SUBSCRIBER_NOT_FOUND;
+  } else {
     SubscriberInfo temp_info;
     temp_info.client_id = to_mqtt_string(client_id, nullptr);
-    subscribers_.erase(temp_info);
+    auto it = subscribers_.find(temp_info);
+    if (it != subscribers_.end()) {
+      subscribers_.erase(it);
+    } else {
+      LOG_DEBUG("Subscriber '{}' not found for removal", client_id);
+      ret = MQ_ERR_TOPIC_TREE_SUBSCRIBER_NOT_FOUND;
+    }
+  }
+
+  return ret;
 }
 
-std::shared_ptr<TopicTreeNode> IntermediateNode::get_child(const std::string& level) const {
+int IntermediateNode::get_child(const std::string& level,
+                                std::shared_ptr<TopicTreeNode>& child) const
+{
+  int ret = MQ_SUCCESS;
+
+  if (level.empty()) {
+    LOG_ERROR("Level string cannot be empty");
+    ret = MQ_ERR_TOPIC_TREE_INVALID_TOPIC;
+  } else {
     auto it = children_.find(level);
-    return (it != children_.end()) ? it->second : nullptr;
+    if (it != children_.end()) {
+      child = it->second;
+    } else {
+      child = nullptr;
+      ret = MQ_ERR_NOT_FOUND_V2;
+    }
+  }
+
+  return ret;
+}
+
+int IntermediateNode::get_children(
+    const TopicTreeMap<std::string, std::shared_ptr<TopicTreeNode>>*& children) const
+{
+  children = &children_;
+  return MQ_SUCCESS;
+}
+
+int IntermediateNode::get_subscribers(const SubscriberSet*& subscribers) const
+{
+  subscribers = &subscribers_;
+  return MQ_SUCCESS;
 }
 
 // 引用计数管理
-void intrusive_ptr_add_ref(IntermediateNode* node) {
-    if (node) {
-        node->ref_count_.fetch_add(1, std::memory_order_relaxed);
-    }
+void intrusive_ptr_add_ref(IntermediateNode* node)
+{
+  if (node) {
+    node->ref_count_.fetch_add(1, std::memory_order_relaxed);
+  }
 }
 
-void intrusive_ptr_release(IntermediateNode* node) {
-    if (node && node->ref_count_.fetch_sub(1, std::memory_order_acq_rel) == 1) {
-        MQTTAllocator* allocator = node->allocator_;
-        node->~IntermediateNode();
-        allocator->deallocate(node, sizeof(IntermediateNode));
-    }
+void intrusive_ptr_release(IntermediateNode* node)
+{
+  if (node && node->ref_count_.fetch_sub(1, std::memory_order_acq_rel) == 1) {
+    MQTTAllocator* allocator = node->allocator_;
+    node->~IntermediateNode();
+    allocator->deallocate(node, sizeof(IntermediateNode));
+  }
 }
 
 // ============================================================================
 // TopicTreeNode 实现
 // ============================================================================
 
-TopicTreeNode::TopicTreeNode(MQTTAllocator* allocator) 
-    : allocator_(allocator) {
-    // 使用allocator分配IntermediateNode
-    void* memory = allocator_->allocate(sizeof(IntermediateNode));
-    if (!memory) {
-        throw std::bad_alloc();
-    }
-    intermediate_node_.store(new(memory) IntermediateNode(allocator_));
+TopicTreeNode::TopicTreeNode(MQTTAllocator* allocator)
+    : allocator_(allocator), intermediate_node_(nullptr)
+{
+  // 构造函数不做内存分配，由init()函数完成
 }
 
-TopicTreeNode::~TopicTreeNode() {
-    IntermediateNode* node = intermediate_node_.load();
-    if (node) {
-        intrusive_ptr_release(node);
-    }
+int TopicTreeNode::init()
+{
+  if (MQ_ISNULL(allocator_)) {
+    LOG_ERROR("Allocator is null");
+    return MQ_ERR_PARAM_V2;
+  }
+
+  if (intermediate_node_.load() != nullptr) {
+    LOG_ERROR("Node already initialized");
+    return MQ_ERR_INVALID_STATE;
+  }
+
+  // 安全分配IntermediateNode
+  void* memory = allocator_->allocate(sizeof(IntermediateNode));
+  if (MQ_ISNULL(memory)) {
+    LOG_ERROR("Failed to allocate memory for IntermediateNode");
+    return MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+  }
+
+  // 使用placement new构造对象（在内存已分配的情况下，placement new不会失败）
+  IntermediateNode* node = new (memory) IntermediateNode(allocator_);
+  // 注意：placement new在有效内存上不会返回nullptr
+
+  intermediate_node_.store(node);
+  return MQ_SUCCESS;
 }
 
-IntermediateNode* TopicTreeNode::get_intermediate_node() const {
-    IntermediateNode* node = intermediate_node_.load(std::memory_order_acquire);
-    intrusive_ptr_add_ref(node);
-    return node;
+TopicTreeNode::~TopicTreeNode()
+{
+  IntermediateNode* node = intermediate_node_.load();
+  if (node) {
+    intrusive_ptr_release(node);
+  }
 }
 
-bool TopicTreeNode::compare_and_swap_intermediate_node(IntermediateNode* expected, IntermediateNode* desired) {
-    intrusive_ptr_add_ref(desired);
-    bool success = intermediate_node_.compare_exchange_weak(expected, desired, std::memory_order_acq_rel);
-    if (success) {
-        intrusive_ptr_release(expected);
+IntermediateNode* TopicTreeNode::get_intermediate_node() const
+{
+  IntermediateNode* node = intermediate_node_.load(std::memory_order_acquire);
+  intrusive_ptr_add_ref(node);
+  return node;
+}
+
+bool TopicTreeNode::compare_and_swap_intermediate_node(IntermediateNode* expected,
+                                                       IntermediateNode* desired)
+{
+  intrusive_ptr_add_ref(desired);
+  bool success =
+      intermediate_node_.compare_exchange_weak(expected, desired, std::memory_order_acq_rel);
+  if (success) {
+    intrusive_ptr_release(expected);
+  } else {
+    intrusive_ptr_release(desired);
+  }
+  return success;
+}
+
+int TopicTreeNode::get_child(const std::string& level, std::shared_ptr<TopicTreeNode>& child) const
+{
+  int ret = MQ_SUCCESS;
+  IntermediateNode* node = nullptr;
+
+  if (level.empty()) {
+    LOG_ERROR("Level string cannot be empty");
+    ret = MQ_ERR_TOPIC_TREE_INVALID_TOPIC;
+  } else {
+    node = get_intermediate_node();
+    if (MQ_ISNULL(node)) {
+      LOG_ERROR("Failed to get intermediate node");
+      ret = MQ_ERR_INTERNAL;
     } else {
-        intrusive_ptr_release(desired);
+      ret = node->get_child(level, child);
+      // 错误已经在IntermediateNode::get_child中处理
     }
-    return success;
+  }
+
+  if (node) {
+    intrusive_ptr_release(node);
+  }
+
+  return ret;
 }
 
-std::shared_ptr<TopicTreeNode> TopicTreeNode::get_child(const std::string& level) const {
-    IntermediateNode* node = get_intermediate_node();
-    std::shared_ptr<TopicTreeNode> result = node->get_child(level);
+int TopicTreeNode::get_children(
+    TopicTreeMap<std::string, std::shared_ptr<TopicTreeNode>>& children) const
+{
+  int ret = MQ_SUCCESS;
+  IntermediateNode* node = nullptr;
+  const TopicTreeMap<std::string, std::shared_ptr<TopicTreeNode>>* children_ptr = nullptr;
+
+  node = get_intermediate_node();
+  if (MQ_ISNULL(node)) {
+    LOG_ERROR("Failed to get intermediate node");
+    ret = MQ_ERR_INTERNAL;
+  } else {
+    ret = node->get_children(children_ptr);
+    if (MQ_SUCC(ret)) {
+      if (MQ_ISNULL(children_ptr)) {
+        LOG_ERROR("Children pointer is null");
+        ret = MQ_ERR_INTERNAL;
+      } else {
+        // Safe copy operation
+        if (children_ptr->empty()) {
+          children.clear();
+        } else {
+          // Pre-check capacity
+          if (children.max_size() < children_ptr->size()) {
+            LOG_ERROR("Children container capacity insufficient");
+            ret = MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+          } else {
+            // Copy elements safely
+            children.clear();
+            for (const auto& child_pair : *children_ptr) {
+              auto result = children.insert(child_pair);
+              if (!result.second) {
+                LOG_ERROR("Failed to copy child node");
+                ret = MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+                break;
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+
+  if (node) {
     intrusive_ptr_release(node);
-    return result;
+  }
+
+  return ret;
 }
 
-TopicTreeMap<std::string, std::shared_ptr<TopicTreeNode>> TopicTreeNode::get_children() const {
-    IntermediateNode* node = get_intermediate_node();
-    TopicTreeMap<std::string, std::shared_ptr<TopicTreeNode>> result = node->get_children();
+int TopicTreeNode::get_subscribers(SubscriberSet& subscribers) const
+{
+  int ret = MQ_SUCCESS;
+  IntermediateNode* node = nullptr;
+  const SubscriberSet* subscribers_ptr = nullptr;
+
+  node = get_intermediate_node();
+  if (MQ_ISNULL(node)) {
+    LOG_ERROR("Failed to get intermediate node");
+    ret = MQ_ERR_INTERNAL;
+  } else {
+    ret = node->get_subscribers(subscribers_ptr);
+    if (MQ_SUCC(ret)) {
+      if (MQ_ISNULL(subscribers_ptr)) {
+        LOG_ERROR("Subscribers pointer is null");
+        ret = MQ_ERR_INTERNAL;
+      } else {
+        // Safe copy operation
+        if (subscribers_ptr->empty()) {
+          subscribers.clear();
+        } else {
+          // Pre-check capacity
+          if (subscribers.max_size() < subscribers_ptr->size()) {
+            LOG_ERROR("Subscribers container capacity insufficient");
+            ret = MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+          } else {
+            // Copy elements safely
+            subscribers.clear();
+            for (const auto& subscriber : *subscribers_ptr) {
+              auto result = subscribers.insert(subscriber);
+              if (!result.second) {
+                LOG_ERROR("Failed to copy subscriber");
+                ret = MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+                break;
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+
+  if (node) {
     intrusive_ptr_release(node);
-    return result;
+  }
+
+  return ret;
 }
 
-SubscriberSet TopicTreeNode::get_subscribers() const {
-    IntermediateNode* node = get_intermediate_node();
-    SubscriberSet result = node->get_subscribers();
+int TopicTreeNode::has_subscribers(bool& has_subs) const
+{
+  int ret = MQ_SUCCESS;
+  IntermediateNode* node = nullptr;
+  const SubscriberSet* subscribers_ptr = nullptr;
+
+  node = get_intermediate_node();
+  if (MQ_ISNULL(node)) {
+    LOG_ERROR("Failed to get intermediate node");
+    ret = MQ_ERR_INTERNAL;
+  } else {
+    ret = node->get_subscribers(subscribers_ptr);
+    if (MQ_SUCC(ret)) {
+      if (MQ_ISNULL(subscribers_ptr)) {
+        LOG_ERROR("Subscribers pointer is null");
+        ret = MQ_ERR_INTERNAL;
+      } else {
+        has_subs = !subscribers_ptr->empty();
+      }
+    }
+  }
+
+  if (node) {
     intrusive_ptr_release(node);
-    return result;
+  }
+
+  return ret;
 }
 
-bool TopicTreeNode::has_subscribers() const {
-    IntermediateNode* node = get_intermediate_node();
-    bool result = !node->get_subscribers().empty();
-    intrusive_ptr_release(node);
-    return result;
-}
+int TopicTreeNode::has_children(bool& has_childs) const
+{
+  int ret = MQ_SUCCESS;
+  IntermediateNode* node = nullptr;
+  const TopicTreeMap<std::string, std::shared_ptr<TopicTreeNode>>* children_ptr = nullptr;
 
-bool TopicTreeNode::has_children() const {
-    IntermediateNode* node = get_intermediate_node();
-    bool result = !node->get_children().empty();
+  node = get_intermediate_node();
+  if (MQ_ISNULL(node)) {
+    LOG_ERROR("Failed to get intermediate node");
+    ret = MQ_ERR_INTERNAL;
+  } else {
+    ret = node->get_children(children_ptr);
+    if (MQ_SUCC(ret)) {
+      if (MQ_ISNULL(children_ptr)) {
+        LOG_ERROR("Children pointer is null");
+        ret = MQ_ERR_INTERNAL;
+      } else {
+        has_childs = !children_ptr->empty();
+      }
+    }
+  }
+
+  if (node) {
     intrusive_ptr_release(node);
-    return result;
+  }
+
+  return ret;
 }
 
 // ============================================================================
 // ConcurrentTopicTree 实现
 // ============================================================================
 
-ConcurrentTopicTree::ConcurrentTopicTree(MQTTAllocator* allocator) 
-    : allocator_(allocator)
-    , root_(std::make_shared<TopicTreeNode>(allocator))
-    , total_subscribers_(0)
-    , total_nodes_(1) {
-}
-
-ConcurrentTopicTree::~ConcurrentTopicTree() {
-}
-
-std::vector<std::string> ConcurrentTopicTree::split_topic_levels(const MQTTString& topic) const {
-    std::vector<std::string> levels;
-    std::string topic_str = from_mqtt_string(topic);
-    
-    if (topic_str.empty()) {
-        return levels;
+ConcurrentTopicTree::ConcurrentTopicTree(MQTTAllocator* allocator)
+    : allocator_(allocator), root_(nullptr), total_subscribers_(0), total_nodes_(0)
+{
+  // 创建并初始化根节点
+  root_ = std::make_shared<TopicTreeNode>(allocator);
+  if (root_) {
+    int ret = root_->init();
+    if (MQ_SUCC(ret)) {
+      total_nodes_.store(1);
+    } else {
+      LOG_ERROR("Failed to initialize root node: {}", ret);
+      root_.reset();
     }
-    
-    size_t start = 0;
-    size_t pos = 0;
-    
-    while ((pos = topic_str.find('/', start)) != std::string::npos) {
-        if (pos > start) {  // 跳过空级别
-            levels.push_back(topic_str.substr(start, pos - start));
+  }
+}
+
+ConcurrentTopicTree::~ConcurrentTopicTree() {}
+
+int ConcurrentTopicTree::split_topic_levels(const MQTTString& topic,
+                                            std::vector<std::string>& levels) const
+{
+  int ret = MQ_SUCCESS;
+
+  levels.clear();
+  std::string topic_str = from_mqtt_string(topic);
+
+  if (topic_str.empty()) {
+    LOG_ERROR("Topic string cannot be empty");
+    ret = MQ_ERR_TOPIC_TREE_INVALID_TOPIC;
+  } else {
+    // Pre-check string length for safety
+    if (topic_str.length() > 65535) {  // MQTT max topic length
+      LOG_ERROR("Topic string too long: {}", topic_str.length());
+      ret = MQ_ERR_TOPIC_TREE_INVALID_TOPIC;
+    } else {
+      size_t start = 0;
+      size_t pos = 0;
+
+      // Reserve space for estimated levels
+      size_t estimated_levels = std::count(topic_str.begin(), topic_str.end(), '/') + 1;
+      if (estimated_levels > 0 && estimated_levels <= 1000) {  // Reasonable limit
+        levels.reserve(estimated_levels);
+      }
+
+      while ((pos = topic_str.find('/', start)) != std::string::npos) {
+        if (pos > start && pos < topic_str.length()) {  // Skip empty levels and check bounds
+          // Safe substring operation
+          if (start <= topic_str.length() && pos <= topic_str.length()) {
+            levels.emplace_back(topic_str.substr(start, pos - start));
+          }
         }
         start = pos + 1;
+      }
+
+      if (start < topic_str.size()) {
+        levels.emplace_back(topic_str.substr(start));
+      }
     }
-    
-    if (start < topic_str.size()) {
-        levels.push_back(topic_str.substr(start));
-    }
-    
-    return levels;
+  }
+
+  return ret;
 }
 
-std::shared_ptr<TopicTreeNode> ConcurrentTopicTree::get_or_create_node(
-    const std::vector<std::string>& levels, 
-    bool create_if_not_exists) {
-    
-    std::shared_ptr<TopicTreeNode> current = root_;
-    
-    for (const std::string& level : levels) {
-        std::shared_ptr<TopicTreeNode> child = current->get_child(level);
-        
-        if (!child) {
-            if (!create_if_not_exists) {
-                return nullptr;
-            }
-            
-            // 创建新子节点
-            child = std::make_shared<TopicTreeNode>(allocator_);
-            
-            // 无锁添加子节点
-            while (true) {
-                IntermediateNode* old_node = current->get_intermediate_node();
-                IntermediateNode* new_node = old_node->clone();
-                new_node->add_child(level, child);
-                
-                if (current->compare_and_swap_intermediate_node(old_node, new_node)) {
-                    intrusive_ptr_release(old_node);
-                    total_nodes_.fetch_add(1, std::memory_order_relaxed);
-                    break;
-                } else {
-                    intrusive_ptr_release(old_node);
-                    // 正确释放使用allocator分配的IntermediateNode
-                    MQTTAllocator* allocator = new_node->get_allocator();
-                    new_node->~IntermediateNode();
-                    allocator->deallocate(new_node, sizeof(IntermediateNode));
-                    // 重试：可能其他线程已经创建了节点
-                    child = current->get_child(level);
-                    if (child) {
-                        break;
-                    }
-                }
-            }
-        }
-        
-        current = child;
-    }
-    
-    return current;
-}
+int ConcurrentTopicTree::get_or_create_node(const std::vector<std::string>& levels,
+                                            bool create_if_not_exists,
+                                            std::shared_ptr<TopicTreeNode>& node)
+{
+  int ret = MQ_SUCCESS;
+  std::shared_ptr<TopicTreeNode> current = root_;
 
-bool ConcurrentTopicTree::add_subscriber_to_node(
-    std::shared_ptr<TopicTreeNode> node, 
-    const SubscriberInfo& subscriber) {
-    
-    while (true) {
-        IntermediateNode* old_node = node->get_intermediate_node();
-        
-        // 检查是否已经存在
-        const auto& existing_subscribers = old_node->get_subscribers();
-        bool already_exists = existing_subscribers.find(subscriber) != existing_subscribers.end();
-        
-        IntermediateNode* new_node = old_node->clone();
-        new_node->add_subscriber(subscriber);
-        
-        if (node->compare_and_swap_intermediate_node(old_node, new_node)) {
-            intrusive_ptr_release(old_node);
-            if (!already_exists) {
-                total_subscribers_.fetch_add(1, std::memory_order_relaxed);
-            }
-            return true;
-        } else {
-            intrusive_ptr_release(old_node);
-            // 正确释放使用allocator分配的IntermediateNode
-            MQTTAllocator* allocator = new_node->get_allocator();
-            new_node->~IntermediateNode();
-            allocator->deallocate(new_node, sizeof(IntermediateNode));
-            // 重试
-        }
-    }
-    
-    return false;
-}
+  for (const std::string& level : levels) {
+    std::shared_ptr<TopicTreeNode> child;
+    ret = current->get_child(level, child);
 
-bool ConcurrentTopicTree::remove_subscriber_from_node(
-    std::shared_ptr<TopicTreeNode> node, 
-    const std::string& client_id) {
-    
-    while (true) {
-        IntermediateNode* old_node = node->get_intermediate_node();
-        
-        // 检查是否存在
-        SubscriberInfo temp_info;
-        temp_info.client_id = to_mqtt_string(client_id, nullptr);
-        const auto& existing_subscribers = old_node->get_subscribers();
-        bool exists = existing_subscribers.find(temp_info) != existing_subscribers.end();
-        
-        if (!exists) {
-            intrusive_ptr_release(old_node);
-            return false;
-        }
-        
-        IntermediateNode* new_node = old_node->clone();
-        new_node->remove_subscriber(client_id);
-        
-        if (node->compare_and_swap_intermediate_node(old_node, new_node)) {
-            intrusive_ptr_release(old_node);
-            total_subscribers_.fetch_sub(1, std::memory_order_relaxed);
-            return true;
-        } else {
-            intrusive_ptr_release(old_node);
-            // 正确释放使用allocator分配的IntermediateNode
-            MQTTAllocator* allocator = new_node->get_allocator();
-            new_node->~IntermediateNode();
-            allocator->deallocate(new_node, sizeof(IntermediateNode));
-            // 重试
-        }
-    }
-    
-    return false;
-}
-
-int ConcurrentTopicTree::subscribe(const MQTTString& topic_filter, const MQTTString& client_id, uint8_t qos) {
-    if (topic_filter.empty() || client_id.empty()) {
-        LOG_ERROR("Invalid topic filter or client ID for subscription");
-        return MQ_ERR_PARAM_V2;
-    }
-    
-    std::vector<std::string> levels = split_topic_levels(topic_filter);
-    if (levels.empty()) {
-        LOG_ERROR("Invalid topic filter format: {}", from_mqtt_string(topic_filter));
-        return MQ_ERR_PARAM_V2;
-    }
-    
-    // MQTT规范：主题过滤器不能以 # 开头，除非是单独的 #
-    std::string topic_str = from_mqtt_string(topic_filter);
-    if (topic_str.find('#') != std::string::npos) {
-        // # 只能出现在末尾，且前面必须是 / 或者是单独的 #
-        size_t hash_pos = topic_str.find('#');
-        if (hash_pos != topic_str.size() - 1 ||  // # 不在末尾
-            (hash_pos > 0 && topic_str[hash_pos - 1] != '/')) {  // # 前面不是 /
-            if (topic_str != "#") {  // 除非是单独的 #
-                LOG_ERROR("Invalid topic filter format: {}", topic_str);
-                return MQ_ERR_PARAM_V2;
-            }
-        }
-    }
-    
-    std::shared_ptr<TopicTreeNode> target_node = get_or_create_node(levels, true);
-    if (!target_node) {
-        LOG_ERROR("Failed to create node for topic filter: {}", from_mqtt_string(topic_filter));
-        return MQ_ERR_INTERNAL;
-    }
-    
-    SubscriberInfo subscriber(client_id, qos);
-    if (!add_subscriber_to_node(target_node, subscriber)) {
-        LOG_ERROR("Failed to add subscriber to node");
-        return MQ_ERR_INTERNAL;
-    }
-    
-    LOG_DEBUG("Subscribed client {} to topic filter {}", from_mqtt_string(client_id), from_mqtt_string(topic_filter));
-    return MQ_SUCCESS;
-}
-
-int ConcurrentTopicTree::unsubscribe(const MQTTString& topic_filter, const MQTTString& client_id) {
-    if (topic_filter.empty() || client_id.empty()) {
-        LOG_ERROR("Invalid topic filter or client ID for unsubscription");
-        return MQ_ERR_PARAM_V2;
-    }
-    
-    std::vector<std::string> levels = split_topic_levels(topic_filter);
-    if (levels.empty()) {
-        LOG_ERROR("Invalid topic filter format: {}", from_mqtt_string(topic_filter));
-        return MQ_ERR_PARAM_V2;
-    }
-    
-    std::shared_ptr<TopicTreeNode> target_node = get_or_create_node(levels, false);
-    if (!target_node) {
-        LOG_WARN("Topic filter not found for unsubscription: {}", from_mqtt_string(topic_filter));
+    if (ret == MQ_ERR_NOT_FOUND_V2) {
+      if (!create_if_not_exists) {
         return MQ_ERR_NOT_FOUND_V2;
+      }
+
+      // Create new child node
+      child = std::make_shared<TopicTreeNode>(allocator_);
+      if (!child) {
+        LOG_ERROR("Failed to create new TopicTreeNode");
+        return MQ_ERR_TOPIC_TREE_MEMORY_ALLOC;
+      }
+
+      // Initialize the new node
+      ret = child->init();
+      if (MQ_FAIL(ret)) {
+        LOG_ERROR("Failed to initialize new TopicTreeNode: {}", ret);
+        return ret;
+      }
+
+      // Lock-free add child node
+      while (true) {
+        IntermediateNode* old_node = current->get_intermediate_node();
+        IntermediateNode* new_node = nullptr;
+        ret = old_node->clone(new_node);
+        if (MQ_FAIL(ret)) {
+          intrusive_ptr_release(old_node);
+          return ret;
+        }
+
+        ret = new_node->add_child(level, child);
+        if (MQ_FAIL(ret)) {
+          intrusive_ptr_release(old_node);
+          MQTTAllocator* allocator = new_node->get_allocator();
+          new_node->~IntermediateNode();
+          allocator->deallocate(new_node, sizeof(IntermediateNode));
+          return ret;
+        }
+
+        if (current->compare_and_swap_intermediate_node(old_node, new_node)) {
+          intrusive_ptr_release(old_node);
+          total_nodes_.fetch_add(1, std::memory_order_relaxed);
+          break;
+        } else {
+          intrusive_ptr_release(old_node);
+          MQTTAllocator* allocator = new_node->get_allocator();
+          new_node->~IntermediateNode();
+          allocator->deallocate(new_node, sizeof(IntermediateNode));
+          // Retry: other thread might have created the node
+          ret = current->get_child(level, child);
+          if (MQ_SUCC(ret)) {
+            break;
+          }
+        }
+      }
+    } else if (MQ_FAIL(ret)) {
+      return ret;
     }
-    
-    std::string client_id_str = from_mqtt_string(client_id);
-    if (!remove_subscriber_from_node(target_node, client_id_str)) {
-        LOG_WARN("Subscriber not found for unsubscription: {} from {}", client_id_str, from_mqtt_string(topic_filter));
-        return MQ_ERR_NOT_FOUND_V2;
-    }
-    
-    LOG_DEBUG("Unsubscribed client {} from topic filter {}", client_id_str, from_mqtt_string(topic_filter));
-    return MQ_SUCCESS;
+
+    current = child;
+  }
+
+  node = current;
+  return ret;
 }
 
-void ConcurrentTopicTree::find_subscribers_recursive(
-    const std::shared_ptr<TopicTreeNode>& node,
-    const std::vector<std::string>& topic_levels,
-    size_t level_index,
-    TopicMatchResult& result) const {
-    
-    if (!node) {
-        return;
+int ConcurrentTopicTree::add_subscriber_to_node(std::shared_ptr<TopicTreeNode> node,
+                                                const SubscriberInfo& subscriber)
+{
+  int ret = MQ_SUCCESS;
+
+  while (true) {
+    IntermediateNode* old_node = node->get_intermediate_node();
+    if (MQ_ISNULL(old_node)) {
+      LOG_ERROR("Failed to get intermediate node");
+      return MQ_ERR_INTERNAL;
     }
-    
-    // 获取当前节点的订阅者
-    auto subscribers = node->get_subscribers();
-    for (const SubscriberInfo& subscriber : subscribers) {
+
+    // Check if already exists
+    const SubscriberSet* existing_subscribers = nullptr;
+    ret = old_node->get_subscribers(existing_subscribers);
+    if (MQ_FAIL(ret)) {
+      intrusive_ptr_release(old_node);
+      return ret;
+    }
+
+    bool already_exists = (existing_subscribers &&
+                           existing_subscribers->find(subscriber) != existing_subscribers->end());
+
+    IntermediateNode* new_node = nullptr;
+    ret = old_node->clone(new_node);
+    if (MQ_FAIL(ret)) {
+      intrusive_ptr_release(old_node);
+      return ret;
+    }
+
+    ret = new_node->add_subscriber(subscriber);
+    if (MQ_FAIL(ret)) {
+      intrusive_ptr_release(old_node);
+      MQTTAllocator* allocator = new_node->get_allocator();
+      new_node->~IntermediateNode();
+      allocator->deallocate(new_node, sizeof(IntermediateNode));
+      return ret;
+    }
+
+    if (node->compare_and_swap_intermediate_node(old_node, new_node)) {
+      intrusive_ptr_release(old_node);
+      if (!already_exists) {
+        total_subscribers_.fetch_add(1, std::memory_order_relaxed);
+      }
+      return MQ_SUCCESS;
+    } else {
+      intrusive_ptr_release(old_node);
+      MQTTAllocator* allocator = new_node->get_allocator();
+      new_node->~IntermediateNode();
+      allocator->deallocate(new_node, sizeof(IntermediateNode));
+      // Retry
+    }
+  }
+
+  return MQ_ERR_INTERNAL;
+}
+
+int ConcurrentTopicTree::remove_subscriber_from_node(std::shared_ptr<TopicTreeNode> node,
+                                                     const std::string& client_id)
+{
+  int ret = MQ_SUCCESS;
+
+  while (true) {
+    IntermediateNode* old_node = node->get_intermediate_node();
+    if (MQ_ISNULL(old_node)) {
+      LOG_ERROR("Failed to get intermediate node");
+      return MQ_ERR_INTERNAL;
+    }
+
+    // Check if exists
+    SubscriberInfo temp_info;
+    temp_info.client_id = to_mqtt_string(client_id, nullptr);
+    const SubscriberSet* existing_subscribers = nullptr;
+    ret = old_node->get_subscribers(existing_subscribers);
+    if (MQ_FAIL(ret)) {
+      intrusive_ptr_release(old_node);
+      return ret;
+    }
+
+    bool exists = (existing_subscribers &&
+                   existing_subscribers->find(temp_info) != existing_subscribers->end());
+
+    if (!exists) {
+      intrusive_ptr_release(old_node);
+      return MQ_ERR_NOT_FOUND_V2;
+    }
+
+    IntermediateNode* new_node = nullptr;
+    ret = old_node->clone(new_node);
+    if (MQ_FAIL(ret)) {
+      intrusive_ptr_release(old_node);
+      return ret;
+    }
+
+    ret = new_node->remove_subscriber(client_id);
+    if (MQ_FAIL(ret)) {
+      intrusive_ptr_release(old_node);
+      MQTTAllocator* allocator = new_node->get_allocator();
+      new_node->~IntermediateNode();
+      allocator->deallocate(new_node, sizeof(IntermediateNode));
+      return ret;
+    }
+
+    if (node->compare_and_swap_intermediate_node(old_node, new_node)) {
+      intrusive_ptr_release(old_node);
+      total_subscribers_.fetch_sub(1, std::memory_order_relaxed);
+      return MQ_SUCCESS;
+    } else {
+      intrusive_ptr_release(old_node);
+      MQTTAllocator* allocator = new_node->get_allocator();
+      new_node->~IntermediateNode();
+      allocator->deallocate(new_node, sizeof(IntermediateNode));
+      // Retry
+    }
+  }
+
+  return MQ_ERR_INTERNAL;
+}
+
+int ConcurrentTopicTree::subscribe(const MQTTString& topic_filter, const MQTTString& client_id,
+                                   uint8_t qos)
+{
+  if (!is_initialized()) {
+    LOG_ERROR("Topic tree not initialized");
+    return MQ_ERR_TOPIC_TREE;
+  }
+
+  if (topic_filter.empty() || client_id.empty()) {
+    LOG_ERROR("Invalid topic filter or client ID for subscription");
+    return MQ_ERR_PARAM_V2;
+  }
+
+  std::vector<std::string> levels;
+  int ret = split_topic_levels(topic_filter, levels);
+  if (MQ_FAIL(ret)) {
+    return ret;
+  }
+
+  if (levels.empty()) {
+    LOG_ERROR("Invalid topic filter format: {}", from_mqtt_string(topic_filter));
+    return MQ_ERR_PARAM_V2;
+  }
+
+  // MQTT规范：主题过滤器不能以 # 开头，除非是单独的 #
+  std::string topic_str = from_mqtt_string(topic_filter);
+  if (topic_str.find('#') != std::string::npos) {
+    // # 只能出现在末尾，且前面必须是 / 或者是单独的 #
+    size_t hash_pos = topic_str.find('#');
+    if (hash_pos != topic_str.size() - 1 ||                  // # 不在末尾
+        (hash_pos > 0 && topic_str[hash_pos - 1] != '/')) {  // # 前面不是 /
+      if (topic_str != "#") {                                // 除非是单独的 #
+        LOG_ERROR("Invalid topic filter format: {}", topic_str);
+        return MQ_ERR_PARAM_V2;
+      }
+    }
+  }
+
+  std::shared_ptr<TopicTreeNode> target_node;
+  ret = get_or_create_node(levels, true, target_node);
+  if (MQ_FAIL(ret)) {
+    LOG_ERROR("Failed to create node for topic filter: {}", from_mqtt_string(topic_filter));
+    return ret;
+  }
+
+  SubscriberInfo subscriber(client_id, qos);
+  ret = add_subscriber_to_node(target_node, subscriber);
+  if (MQ_FAIL(ret)) {
+    LOG_ERROR("Failed to add subscriber to node");
+    return ret;
+  }
+
+  LOG_DEBUG("Subscribed client {} to topic filter {}", from_mqtt_string(client_id),
+            from_mqtt_string(topic_filter));
+  return MQ_SUCCESS;
+}
+
+int ConcurrentTopicTree::unsubscribe(const MQTTString& topic_filter, const MQTTString& client_id)
+{
+  if (!is_initialized()) {
+    LOG_ERROR("Topic tree not initialized");
+    return MQ_ERR_TOPIC_TREE;
+  }
+
+  if (topic_filter.empty() || client_id.empty()) {
+    LOG_ERROR("Invalid topic filter or client ID for unsubscription");
+    return MQ_ERR_PARAM_V2;
+  }
+
+  std::vector<std::string> levels;
+  int ret = split_topic_levels(topic_filter, levels);
+  if (MQ_FAIL(ret)) {
+    return ret;
+  }
+
+  if (levels.empty()) {
+    LOG_ERROR("Invalid topic filter format: {}", from_mqtt_string(topic_filter));
+    return MQ_ERR_PARAM_V2;
+  }
+
+  std::shared_ptr<TopicTreeNode> target_node;
+  ret = get_or_create_node(levels, false, target_node);
+  if (MQ_FAIL(ret)) {
+    LOG_WARN("Topic filter not found for unsubscription: {}", from_mqtt_string(topic_filter));
+    return ret;
+  }
+
+  std::string client_id_str = from_mqtt_string(client_id);
+  ret = remove_subscriber_from_node(target_node, client_id_str);
+  if (MQ_FAIL(ret)) {
+    LOG_WARN("Subscriber not found for unsubscription: {} from {}", client_id_str,
+             from_mqtt_string(topic_filter));
+    return ret;
+  }
+
+  LOG_DEBUG("Unsubscribed client {} from topic filter {}", client_id_str,
+            from_mqtt_string(topic_filter));
+  return MQ_SUCCESS;
+}
+
+int ConcurrentTopicTree::find_subscribers_recursive(const std::shared_ptr<TopicTreeNode>& node,
+                                                    const std::vector<std::string>& topic_levels,
+                                                    size_t level_index,
+                                                    TopicMatchResult& result) const
+{
+  if (!node) {
+    return MQ_SUCCESS;
+  }
+
+  int ret = MQ_SUCCESS;
+
+  // 获取当前节点的订阅者
+  SubscriberSet subscribers;
+  ret = node->get_subscribers(subscribers);
+  if (MQ_FAIL(ret)) {
+    return ret;
+  }
+
+  for (const SubscriberInfo& subscriber : subscribers) {
+    result.subscribers.push_back(subscriber);
+    result.total_count++;
+  }
+
+  // 如果已经到达主题末尾，检查多级通配符
+  if (level_index >= topic_levels.size()) {
+    std::shared_ptr<TopicTreeNode> wildcard_node;
+    ret = node->get_child("#", wildcard_node);
+    if (ret == MQ_SUCCESS && wildcard_node) {
+      SubscriberSet wildcard_subscribers;
+      ret = wildcard_node->get_subscribers(wildcard_subscribers);
+      if (MQ_SUCC(ret)) {
+        for (const SubscriberInfo& subscriber : wildcard_subscribers) {
+          result.subscribers.push_back(subscriber);
+          result.total_count++;
+        }
+      }
+    }
+    return MQ_SUCCESS;
+  }
+
+  const std::string& current_level = topic_levels[level_index];
+
+  // 1. 精确匹配
+  std::shared_ptr<TopicTreeNode> exact_child;
+  ret = node->get_child(current_level, exact_child);
+  if (ret == MQ_SUCCESS && exact_child) {
+    ret = find_subscribers_recursive(exact_child, topic_levels, level_index + 1, result);
+    if (MQ_FAIL(ret)) {
+      return ret;
+    }
+  }
+
+  // 2. 单级通配符 '+'
+  std::shared_ptr<TopicTreeNode> single_wildcard_child;
+  ret = node->get_child("+", single_wildcard_child);
+  if (ret == MQ_SUCCESS && single_wildcard_child) {
+    ret = find_subscribers_recursive(single_wildcard_child, topic_levels, level_index + 1, result);
+    if (MQ_FAIL(ret)) {
+      return ret;
+    }
+  }
+
+  // 3. 多级通配符 '#' (匹配当前级别及之后的所有级别)
+  std::shared_ptr<TopicTreeNode> multi_wildcard_child;
+  ret = node->get_child("#", multi_wildcard_child);
+  if (ret == MQ_SUCCESS && multi_wildcard_child) {
+    SubscriberSet multi_wildcard_subscribers;
+    ret = multi_wildcard_child->get_subscribers(multi_wildcard_subscribers);
+    if (MQ_SUCC(ret)) {
+      for (const SubscriberInfo& subscriber : multi_wildcard_subscribers) {
         result.subscribers.push_back(subscriber);
         result.total_count++;
+      }
     }
-    
-    // 如果已经到达主题末尾，检查多级通配符
-    if (level_index >= topic_levels.size()) {
-        std::shared_ptr<TopicTreeNode> wildcard_node = node->get_child("#");
-        if (wildcard_node) {
-            auto wildcard_subscribers = wildcard_node->get_subscribers();
-            for (const SubscriberInfo& subscriber : wildcard_subscribers) {
-                result.subscribers.push_back(subscriber);
-                result.total_count++;
-            }
-        }
-        return;
-    }
-    
-    const std::string& current_level = topic_levels[level_index];
-    
-    // 1. 精确匹配
-    std::shared_ptr<TopicTreeNode> exact_child = node->get_child(current_level);
-    if (exact_child) {
-        find_subscribers_recursive(exact_child, topic_levels, level_index + 1, result);
-    }
-    
-    // 2. 单级通配符 '+'
-    std::shared_ptr<TopicTreeNode> single_wildcard_child = node->get_child("+");
-    if (single_wildcard_child) {
-        find_subscribers_recursive(single_wildcard_child, topic_levels, level_index + 1, result);
-    }
-    
-    // 3. 多级通配符 '#' (匹配当前级别及之后的所有级别)
-    std::shared_ptr<TopicTreeNode> multi_wildcard_child = node->get_child("#");
-    if (multi_wildcard_child) {
-        auto multi_wildcard_subscribers = multi_wildcard_child->get_subscribers();
-        for (const SubscriberInfo& subscriber : multi_wildcard_subscribers) {
-            result.subscribers.push_back(subscriber);
-            result.total_count++;
-        }
-    }
+  }
+
+  return MQ_SUCCESS;
 }
 
-TopicMatchResult ConcurrentTopicTree::find_subscribers(const MQTTString& topic) const {
-    TopicMatchResult result(allocator_);
-    
-    if (topic.empty()) {
-        LOG_WARN("Empty topic for subscriber search");
-        return result;
-    }
-    
-    std::vector<std::string> levels = split_topic_levels(topic);
-    
-    // 从根节点开始递归查找
-    find_subscribers_recursive(root_, levels, 0, result);
-    
-    // 去重（可能有重复的订阅者）
-    std::sort(result.subscribers.begin(), result.subscribers.end(), 
+int ConcurrentTopicTree::find_subscribers(const MQTTString& topic, TopicMatchResult& result) const
+{
+  if (!is_initialized()) {
+    LOG_ERROR("Topic tree not initialized");
+    return MQ_ERR_TOPIC_TREE;
+  }
+
+  if (topic.empty()) {
+    LOG_WARN("Empty topic for subscriber search");
+    return MQ_ERR_PARAM_V2;
+  }
+
+  std::vector<std::string> levels;
+  int ret = split_topic_levels(topic, levels);
+  if (MQ_FAIL(ret)) {
+    return ret;
+  }
+
+  // 从根节点开始递归查找
+  ret = find_subscribers_recursive(root_, levels, 0, result);
+  if (MQ_FAIL(ret)) {
+    return ret;
+  }
+
+  // 去重（可能有重复的订阅者）
+  if (!result.subscribers.empty()) {
+    std::sort(result.subscribers.begin(), result.subscribers.end(),
               [](const SubscriberInfo& a, const SubscriberInfo& b) {
-                  return from_mqtt_string(a.client_id) < from_mqtt_string(b.client_id);
+                return from_mqtt_string(a.client_id) < from_mqtt_string(b.client_id);
               });
-    
-    result.subscribers.erase(
-        std::unique(result.subscribers.begin(), result.subscribers.end()),
-        result.subscribers.end());
-    
+
+    result.subscribers.erase(std::unique(result.subscribers.begin(), result.subscribers.end()),
+                             result.subscribers.end());
+
     result.total_count = result.subscribers.size();
-    
-    LOG_DEBUG("Found {} subscribers for topic: {}", result.total_count, from_mqtt_string(topic));
-    return result;
+  }
+
+  LOG_DEBUG("Found {} subscribers for topic: {}", result.total_count, from_mqtt_string(topic));
+  return MQ_SUCCESS;
 }
 
-int ConcurrentTopicTree::unsubscribe_all(const MQTTString& client_id) {
-    if (client_id.empty()) {
-        LOG_ERROR("Invalid client ID for unsubscribe_all");
-        return 0;
-    }
-    
-    // 获取客户端的所有订阅
-    std::vector<MQTTString> subscriptions = get_client_subscriptions(client_id);
-    
-    int unsubscribed_count = 0;
-    for (const MQTTString& subscription : subscriptions) {
-        if (unsubscribe(subscription, client_id) == MQ_SUCCESS) {
-            unsubscribed_count++;
-        }
-    }
-    
-    LOG_DEBUG("Unsubscribed client {} from {} topic filters", 
-              from_mqtt_string(client_id), unsubscribed_count);
-    return unsubscribed_count;
-}
-
-void ConcurrentTopicTree::collect_client_subscriptions(
-    const std::shared_ptr<TopicTreeNode>& node,
-    const std::string& current_path,
-    const std::string& client_id,
-    std::vector<MQTTString>& result) const {
-    
-    if (!node) {
-        return;
-    }
-    
-    // 检查当前节点是否有目标客户端的订阅
-    auto subscribers = node->get_subscribers();
-    SubscriberInfo target_info;
-    target_info.client_id = to_mqtt_string(client_id, nullptr);
-    
-    if (subscribers.find(target_info) != subscribers.end()) {
-        result.push_back(to_mqtt_string(current_path, nullptr));
-    }
-    
-    // 递归检查子节点
-    auto children = node->get_children();
-    for (const auto& child_pair : children) {
-        std::string child_path = current_path.empty() ? child_pair.first : (current_path + "/" + child_pair.first);
-        collect_client_subscriptions(child_pair.second, child_path, client_id, result);
-    }
-}
-
-std::vector<MQTTString> ConcurrentTopicTree::get_client_subscriptions(const MQTTString& client_id) const {
-    std::vector<MQTTString> result;
-    std::string client_id_str = from_mqtt_string(client_id);
-    collect_client_subscriptions(root_, "", client_id_str, result);
-    return result;
-}
-
-size_t ConcurrentTopicTree::get_total_subscribers() const {
-    return total_subscribers_.load(std::memory_order_relaxed);
-}
-
-size_t ConcurrentTopicTree::get_total_nodes() const {
-    return total_nodes_.load(std::memory_order_relaxed);
-}
-
-bool ConcurrentTopicTree::cleanup_empty_nodes_recursive(std::shared_ptr<TopicTreeNode> node, const std::string& path) {
-    if (!node) {
-        return true;
-    }
-    
-    auto children = node->get_children();
-    
-    // 递归清理子节点
-    std::vector<std::string> children_to_remove;
-    for (const auto& child_pair : children) {
-        std::string child_path = path.empty() ? child_pair.first : (path + "/" + child_pair.first);
-        if (cleanup_empty_nodes_recursive(child_pair.second, child_path)) {
-            children_to_remove.push_back(child_pair.first);
-        }
-    }
-    
-    // 移除空的子节点
-    for (const std::string& child_level : children_to_remove) {
-        while (true) {
-            IntermediateNode* old_node = node->get_intermediate_node();
-            IntermediateNode* new_node = old_node->clone();
-            new_node->remove_child(child_level);
-            
-            if (node->compare_and_swap_intermediate_node(old_node, new_node)) {
-                intrusive_ptr_release(old_node);
-                total_nodes_.fetch_sub(1, std::memory_order_relaxed);
-                break;
-            } else {
-                intrusive_ptr_release(old_node);
-                // 正确释放使用allocator分配的IntermediateNode
-                MQTTAllocator* allocator = new_node->get_allocator();
-                new_node->~IntermediateNode();
-                allocator->deallocate(new_node, sizeof(IntermediateNode));
-                // 重试
-            }
-        }
-    }
-    
-    // 判断当前节点是否应该被删除（没有订阅者且没有子节点）
-    return !node->has_subscribers() && !node->has_children();
-}
-
-size_t ConcurrentTopicTree::cleanup_empty_nodes() {
-    size_t original_count = total_nodes_.load(std::memory_order_relaxed);
-    cleanup_empty_nodes_recursive(root_, "");
-    size_t final_count = total_nodes_.load(std::memory_order_relaxed);
-    size_t cleaned_count = original_count - final_count;
-    
-    if (cleaned_count > 0) {
-        LOG_DEBUG("Cleaned up {} empty nodes", cleaned_count);
-    }
-    
-    return cleaned_count;
-}
-
-size_t ConcurrentTopicTree::get_memory_usage() const {
-    if (allocator_) {
-        return allocator_->get_memory_usage();
-    }
+int ConcurrentTopicTree::unsubscribe_all(const MQTTString& client_id)
+{
+  if (client_id.empty()) {
+    LOG_ERROR("Invalid client ID for unsubscribe_all");
     return 0;
+  }
+
+  // 获取客户端的所有订阅
+  std::vector<MQTTString> subscriptions;
+  int ret = get_client_subscriptions(client_id, subscriptions);
+  if (MQ_FAIL(ret)) {
+    LOG_ERROR("Failed to get client subscriptions");
+    return 0;
+  }
+
+  int unsubscribed_count = 0;
+  for (const MQTTString& subscription : subscriptions) {
+    if (unsubscribe(subscription, client_id) == MQ_SUCCESS) {
+      unsubscribed_count++;
+    }
+  }
+
+  LOG_DEBUG("Unsubscribed client {} from {} topic filters", from_mqtt_string(client_id),
+            unsubscribed_count);
+  return unsubscribed_count;
+}
+
+int ConcurrentTopicTree::collect_client_subscriptions(const std::shared_ptr<TopicTreeNode>& node,
+                                                      const std::string& current_path,
+                                                      const std::string& client_id,
+                                                      std::vector<MQTTString>& result) const
+{
+  if (!node) {
+    return MQ_SUCCESS;
+  }
+
+  int ret = MQ_SUCCESS;
+
+  // 检查当前节点是否有目标客户端的订阅
+  SubscriberSet subscribers;
+  ret = node->get_subscribers(subscribers);
+  if (MQ_FAIL(ret)) {
+    return ret;
+  }
+
+  SubscriberInfo target_info;
+  target_info.client_id = to_mqtt_string(client_id, nullptr);
+
+  if (subscribers.find(target_info) != subscribers.end()) {
+    result.push_back(to_mqtt_string(current_path, nullptr));
+  }
+
+  // 递归检查子节点
+  TopicTreeMap<std::string, std::shared_ptr<TopicTreeNode>> children;
+  ret = node->get_children(children);
+  if (MQ_FAIL(ret)) {
+    return ret;
+  }
+
+  for (const auto& child_pair : children) {
+    std::string child_path =
+        current_path.empty() ? child_pair.first : (current_path + "/" + child_pair.first);
+    ret = collect_client_subscriptions(child_pair.second, child_path, client_id, result);
+    if (MQ_FAIL(ret)) {
+      return ret;
+    }
+  }
+
+  return MQ_SUCCESS;
+}
+
+int ConcurrentTopicTree::get_client_subscriptions(const MQTTString& client_id,
+                                                  std::vector<MQTTString>& subscriptions) const
+{
+  subscriptions.clear();
+
+  if (!is_initialized()) {
+    LOG_ERROR("Topic tree not initialized");
+    return MQ_ERR_TOPIC_TREE;
+  }
+
+  std::string client_id_str = from_mqtt_string(client_id);
+  return collect_client_subscriptions(root_, "", client_id_str, subscriptions);
+}
+
+int ConcurrentTopicTree::get_total_subscribers(size_t& total_count) const
+{
+  total_count = total_subscribers_.load(std::memory_order_relaxed);
+  return MQ_SUCCESS;
+}
+
+int ConcurrentTopicTree::get_total_nodes(size_t& total_count) const
+{
+  total_count = total_nodes_.load(std::memory_order_relaxed);
+  return MQ_SUCCESS;
+}
+
+int ConcurrentTopicTree::cleanup_empty_nodes_recursive(std::shared_ptr<TopicTreeNode> node,
+                                                       const std::string& path, bool& should_delete)
+{
+  should_delete = false;
+
+  if (!node) {
+    should_delete = true;
+    return MQ_SUCCESS;
+  }
+
+  int ret = MQ_SUCCESS;
+  TopicTreeMap<std::string, std::shared_ptr<TopicTreeNode>> children;
+  ret = node->get_children(children);
+  if (MQ_FAIL(ret)) {
+    return ret;
+  }
+
+  // 递归清理子节点
+  std::vector<std::string> children_to_remove;
+  for (const auto& child_pair : children) {
+    std::string child_path = path.empty() ? child_pair.first : (path + "/" + child_pair.first);
+    bool child_should_delete = false;
+    ret = cleanup_empty_nodes_recursive(child_pair.second, child_path, child_should_delete);
+    if (MQ_FAIL(ret)) {
+      return ret;
+    }
+    if (child_should_delete) {
+      children_to_remove.push_back(child_pair.first);
+    }
+  }
+
+  // 移除空的子节点
+  for (const std::string& child_level : children_to_remove) {
+    while (true) {
+      IntermediateNode* old_node = node->get_intermediate_node();
+      if (MQ_ISNULL(old_node)) {
+        LOG_ERROR("Failed to get intermediate node");
+        return MQ_ERR_INTERNAL;
+      }
+
+      IntermediateNode* new_node = nullptr;
+      ret = old_node->clone(new_node);
+      if (MQ_FAIL(ret)) {
+        intrusive_ptr_release(old_node);
+        return ret;
+      }
+
+      ret = new_node->remove_child(child_level);
+      if (MQ_FAIL(ret)) {
+        intrusive_ptr_release(old_node);
+        MQTTAllocator* allocator = new_node->get_allocator();
+        new_node->~IntermediateNode();
+        allocator->deallocate(new_node, sizeof(IntermediateNode));
+        return ret;
+      }
+
+      if (node->compare_and_swap_intermediate_node(old_node, new_node)) {
+        intrusive_ptr_release(old_node);
+        total_nodes_.fetch_sub(1, std::memory_order_relaxed);
+        break;
+      } else {
+        intrusive_ptr_release(old_node);
+        MQTTAllocator* allocator = new_node->get_allocator();
+        new_node->~IntermediateNode();
+        allocator->deallocate(new_node, sizeof(IntermediateNode));
+        // 重试
+      }
+    }
+  }
+
+  // 判断当前节点是否应该被删除（没有订阅者且没有子节点）
+  bool has_subs = false;
+  bool has_childs = false;
+  ret = node->has_subscribers(has_subs);
+  if (MQ_FAIL(ret)) {
+    return ret;
+  }
+  ret = node->has_children(has_childs);
+  if (MQ_FAIL(ret)) {
+    return ret;
+  }
+
+  should_delete = !has_subs && !has_childs;
+  return MQ_SUCCESS;
+}
+
+int ConcurrentTopicTree::cleanup_empty_nodes(size_t& cleaned_count)
+{
+  size_t original_count = total_nodes_.load(std::memory_order_relaxed);
+  bool should_delete = false;
+  int ret = cleanup_empty_nodes_recursive(root_, "", should_delete);
+  if (MQ_FAIL(ret)) {
+    return ret;
+  }
+
+  size_t final_count = total_nodes_.load(std::memory_order_relaxed);
+  cleaned_count = original_count - final_count;
+
+  if (cleaned_count > 0) {
+    LOG_DEBUG("Cleaned up {} empty nodes", cleaned_count);
+  }
+
+  return MQ_SUCCESS;
+}
+
+int ConcurrentTopicTree::get_memory_usage(size_t& memory_usage) const
+{
+  if (allocator_) {
+    memory_usage = allocator_->get_memory_usage();
+  } else {
+    memory_usage = 0;
+  }
+  return MQ_SUCCESS;
 }
 
 }  // namespace mqtt
